Remove categorizable-model mixin
Un-DRY these models up to make them easier to convert to ES classes and
octane patterns.
