+++ addon/models/program-year.js
@@ -1,11 +1,11 @@
 import Model, { hasMany, belongsTo, attr } from '@ember-data/model';
 import { computed } from '@ember/object';
-import CategorizableModel from 'ilios-common/mixins/categorizable-model';
 import sortableByPosition from 'ilios-common/utils/sortable-by-position';
+import { all } from 'rsvp';
 
 const { alias, oneWay, collect, sum, not } = computed;
 
-export default Model.extend(CategorizableModel, {
+export default Model.extend({
   startYear: attr('string'),
   locked: attr('boolean'),
   archived: attr('boolean'),
@@ -17,6 +17,7 @@ export default Model.extend(CategorizableModel, {
   competencies: hasMany('competency', {async: true}),
   objectives: hasMany('objective', {async: true}),
   stewards: hasMany('program-year-steward', {async: true}),
+  terms: hasMany('term', {async: true}),
 
   assignableVocabularies: alias('program.school.vocabularies'),
 
@@ -49,6 +50,44 @@ export default Model.extend(CategorizableModel, {
   sortedObjectives: computed('objectives.@each.position', 'objectives.@each.title', async function() {
     const objectives = await this.get('objectives');
     return objectives.toArray().sort(sortableByPosition);
+  }),
+
+  /**
+   * A list of all vocabularies that are associated via terms.
+   * @property associatedVocabularies
+   * @type {Ember.computed}
+   * @public
+   */
+  associatedVocabularies: computed('terms.@each.vocabulary', async function () {
+    const terms = await this.get('terms');
+    const vocabularies = await all(terms.toArray().mapBy('vocabulary'));
+    return vocabularies.uniq().sortBy('title');
+  }),
+
+  /**
+   * A list containing all associated terms and their parent terms.
+   * @property termsWithAllParents
+   * @type {Ember.computed}
+   * @public
+   */
+  termsWithAllParents: computed('terms.[]', async function () {
+    const terms = await this.get('terms');
+    const allTerms = await all(terms.toArray().mapBy('termWithAllParents'));
+    return (allTerms.reduce((array, set) => {
+      array.pushObjects(set);
+      return array;
+    }, [])).uniq();
+  }),
+
+  /**
+   * The number of terms attached to this model
+   * @property termCount
+   * @type {Ember.computed}
+   * @public
+   */
+  termCount: computed('terms.[]', function(){
+    const termIds = this.hasMany('terms').ids();
+    return termIds.length;
   }),
 
   init() {
